+++ odoo/local-src/specific_sale/models/sale.py
@@ -165,18 +165,6 @@ class SaleOrderLine(models.Model):
                         break
             line.exception = exception
 
-    product_qty_remains_to_deliver = fields.Float(
-        string='Remains to deliver',
-        digits=dp.get_precision('Product Unit of Measure'),
-        compute='_compute_product_qty_remains_to_deliver',
-    )
-
-    def _compute_product_qty_remains_to_deliver(self):
-        for line in self:
-            line.product_qty_remains_to_deliver = (
-                line.product_uom_qty - line.qty_delivered
-            )
-
     product_qty_unavailable = fields.Float(
         string='Quantity unavailable',
         digits=dp.get_precision('Product Unit of Measure'),